Take into account categories going over in income
